Switched from Double to Int -.-

+++ src/main/scala/Action.scala
@@ -1,12 +1,12 @@
 package com.somunia.battle
 
-class Action(val effect: Effect, val progress: Double) {
-    def distanceTo(targetProgress: Double): Double = {
+class Action(val effect: Effect, val progress: Int) {
+    def distanceTo(targetProgress: Int): Int = {
         if (targetProgress <= this.toEffect) return this.toEffect - targetProgress
         if (targetProgress <= this.toSelect) return this.toSelect - targetProgress
         0
     }
 
-    val toEffect: Double = progress + effect.preparation
-    val toSelect: Double = progress + effect.preparation + effect.cooldown
+    val toEffect: Int = progress + effect.preparation
+    val toSelect: Int = progress + effect.preparation + effect.cooldown
 }
