+++ src/main/scala/Competitor.scala
@@ -1,5 +1,5 @@
 package com.somunia.battle
 
 class Competitor(var action: Action) {
-    def distanceTo(targetProgress: Double): Double = action.distanceTo(targetProgress)
+    def distanceTo(targetProgress: Int): Int = action.distanceTo(targetProgress)
 }
